+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+===== v0.1.0
+
++Add transfer action
++Add withdraw action
++Add build action
++Add T2 workers
++Rework worker priorities
 *Refactor spawn.manager configruration 
 
 ===== v0.0.1